Update README.md
Hello :) 
Your code is a great help! I am learning a lot from it. Thank you very much for it!

I only had some difficulties finding out where the EventHandler-Code belongs to. I searched in the cakephp4 CookBook and found it out after a while.
So i thought it would be easier for others, if your README text mentions it.

Best Regards